# api/openapi.yaml
openapi: 3.0.3
info:
  title: Darkory
  version: 0.1.0
  description: |
    The canonical contract between Members and a Darkory Install. The CLI, the MCP server and the
    web app use only these operations; humans and agents call the same ones.

    **Versioning.** Paths are versioned (`/v1`). Changes within `/v1` are additive only.

    **Credentials.** Every operation except `getHealth`, `showLoginLink`, `redeemLoginLink` and
    `requestEmailSignIn` needs a credential: a bearer token (`dk_…`) together with a
    `Darkory-Session` header naming the Session making the request, or the browser cookie
    `darkory_session`. A bearer request without `Darkory-Session` is refused with
    `session_required`. A request with no credential is never a Member. A browser Session ends
    after a time unused and after a time in all (Install settings); a deactivated Member's
    credentials are all refused.

    **References.** Path parameters and request-body references take an id, or the natural key
    where one exists: a Task or Feature display key (`WEB-42`), a Team key (`WEB`), a Skill name
    (`qa`), a Member name. Tokens, Evidence and Sessions are referenced by id only. Responses
    always carry ids, plus display keys where they exist.

    **Retries.** Every write accepts an `Idempotency-Key` header. The first response is kept for
    24 hours per Member and key; a retry with the same key returns it unchanged. Reusing a key
...

...

...
    omitted.
  license:
    name: Apache-2.0
...
servers:
...
security:
  - bearerAuth: []
    sessionHeader: []
  - cookieAuth: []
...
paths:
...
  /v1/features/{feature}/evidence:
    post:
      operationId: attachFeatureEvidence
      tags: [evidence]
      summary: Attach Evidence to a Feature
      description: |
        The request body is the file itself, sent with its own `Content-Type` and a
        `Content-Length`; the Install's limit is 100 MiB unless set otherwise. By the Feature's
        owner or a Member of its Team. Errors: `forbidden`, `too_large`.
      parameters:
        - {$ref: "#/components/parameters/FeatureRef"}
        - {$ref: "#/components/parameters/EvidenceFilename"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          "*/*":
            schema: {type: string, format: binary}
      responses:
        "201":
          description: The Evidence record.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Evidence"}
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- tasks
...
  /v1/tasks/{task}/evidence:
    post:
      operationId: attachTaskEvidence
      tags: [evidence]
      summary: Attach Evidence to a Task
      description: |
        The request body is the file itself, sent with its own `Content-Type` and a
        `Content-Length`; the Install's limit is 100 MiB unless set otherwise. Needs the Task's
        Claim while it is held, else its Feature's ownership or membership of its Team. Errors:
        `not_holder`, `forbidden`, `too_large`.
      parameters:
        - {$ref: "#/components/parameters/TaskRef"}
        - {$ref: "#/components/parameters/EvidenceFilename"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          "*/*":
            schema: {type: string, format: binary}
      responses:
        "201":
          description: The Evidence record.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Evidence"}
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- evidence
  /v1/evidence/{evidence}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get an Evidence record
      parameters:
        - {$ref: "#/components/parameters/EvidenceID"}
      responses:
        "200":
          description: The Evidence record.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Evidence"}
        default: {$ref: "#/components/responses/Error"}
  /v1/evidence/{evidence}/content:
    get:
      operationId: downloadEvidence
      tags: [evidence]
      summary: Download an Evidence file
      description: |
        Served with the content type it was attached with, always as an attachment and with
        `X-Content-Type-Options: nosniff`, so a browser never renders an uploaded page.
      parameters:
        - {$ref: "#/components/parameters/EvidenceID"}
      responses:
        "200":
          description: The file.
          headers:
            Content-Disposition:
              description: "`attachment`, with the file's name."
              schema: {type: string}
            X-Content-Type-Options:
              schema: {type: string, enum: [nosniff]}
          content:
            "*/*":
              schema: {type: string, format: binary}
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- activity
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A Member token, `dk_…`. Sent together with `Darkory-Session`.
    sessionHeader:
      type: apiKey
      in: header
      name: Darkory-Session
      description: The id the running copy chose for its Session. Required with a bearer token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: darkory_session
      description: The browser Session, set by redeeming a login link.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        A key unique to this write. A retry with the same key returns the first response. It is 1
        to 255 printable ASCII characters, without spaces; any other is refused with `invalid`.
      schema: {type: string, minLength: 1, maxLength: 255, pattern: "^[!-~]+$"}
...
    FeatureRef:
      name: feature
      in: path
      required: true
      description: Feature id or display key, such as `WEB-1`.
      schema: {type: string}
    TaskRef:
      name: task
      in: path
      required: true
      description: Task id or display key, such as `WEB-42`.
      schema: {type: string}
...
    EvidenceID:
      name: evidence
      in: path
      required: true
      schema: {type: string}
...
    EvidenceFilename:
      name: filename
      in: query
      required: true
      description: The file's name, as it should be shown and downloaded.
      schema: {type: string, minLength: 1, maxLength: 255}
...
  responses:
    Error:
      description: The request failed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: {$ref: "#/components/schemas/ErrorCode"}
        message:
          type: string
          description: For people; may change between releases.
        details:
          type: object
          additionalProperties: true
          description: Extra facts about the failure, by code.
    ErrorCode:
      type: string
      description: |
        Stable reason for a failure, with its HTTP status:
        `invalid` 400 · `unauthenticated` 401 · `session_required` 401 · `forbidden` 403 ·
        `not_found` 404 · `conflict` 409 · `already_claimed` 409 · `not_takeable` 409 ·
        `not_holder` 409 · `ended` 409 · `tasks_open` 409 · `cycle` 409 · `proposal_stale` 409 ·
        `too_large` 413 · `idempotency_key_reused` 422 · `too_many_requests` 429 · `internal` 500 ·
        `not_implemented` 501.
      enum:
        - invalid
        - unauthenticated
        - session_required
        - forbidden
        - not_found
        - conflict
        - already_claimed
        - not_takeable
        - not_holder
        - ended
        - tasks_open
        - cycle
        - proposal_stale
        - too_large
        - idempotency_key_reused
        - too_many_requests
        - internal
        - not_implemented

    # ------------------------------------------------------------ install
...
    Evidence:
      type: object
      required: [id, feature_id, filename, content_type, size, sha256, attached_by, created_at]
      properties:
        id: {type: string}
        feature_id: {type: string}
        task_id:
          type: string
          description: Absent when attached to the Feature itself.
        filename: {type: string}
        content_type: {type: string}
        size:
          type: integer
          format: int64
        sha256: {type: string}
        attached_by: {type: string}
        created_at: {type: string, format: date-time}

    # ------------------------------------------------------------ activity
...